refactor(direct-control): extract setup constants
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only setup/lifecycle command strings, UI loading-state values, and setup parameter IDs from packages/civ7-direct-control/src/index.ts into packages/civ7-direct-control/src/setup/constants.ts while keeping public facade re-exports in index.ts.

Preserves restart, begin, exit-to-main-menu, reload-UI command text, loading-state numeric values, setup parameter IDs, and existing setup/lifecycle facade injection behavior. Broader public constants/types, procedure schemas, operation/proof telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance remain pending.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- public-api.test.ts setup-and-lifecycle.test.ts restart-lifecycle.test.ts session.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership rg scans

Local package/constants relocation proof only; no runtime/live-game, hotseat product-path, AI-ingestion, semantic CLI, telemetry, or procedure-core proof claimed.
